CHANGE INDEX TYPES Ix, Ixs TO usize, isize
This is important for iterator efficiency (outside of the already
efficient contiguous cases). Counted and indexed loops compile much
better using machine-size indices, so for example strided inner
axes are iterated much quicker.

It also allows bigger arrays (bigger axis dimension and bigger strides).

Arrays and views become larger in size, but the difference should be
small. In the bigger data use cases, this difference should be
insignificant and the iteration wins significant.

[breaking-change]

+++ src/lib.rs
@@ -110,9 +110,9 @@ mod shape_error;
 // NOTE: In theory, the whole library should compile
 // and pass tests even if you change Ix and Ixs.
 /// Array index type
-pub type Ix = u32;
+pub type Ix = usize;
 /// Array index type (signed)
-pub type Ixs = i32;
+pub type Ixs = isize;
 
 /// An *N*-dimensional array.
 ///
